PUBAPI-641: Move forward if a machine billing_id does not belong to any existing package.
This should not happen, in theory. But it is perfectly possible to set machine.billing_id to whatever fancy thing and, even in the
case it is being set to an UUID, nobody will grant that the given UUID matches a package.uuid. Therefore, just preventing error and
setting package_name to empty string for these cases.